# swagger.yaml
openapi: 3.0.1
info:
  title: App reviews analysis API
  description: API per l'extracció d'emocions i features a partir d'un text
  version: 1.0.0
servers:
- url: http://localhost:3000/
paths:
  /extract-emotion:
    post:
      summary: Extreure l'emoció d'un conjunt de textos
      description: Retorna un json que conté l'emoció o emocions extretes a partir
        d'un text
      parameters:
      - name: model_emotion
        in: query
        description: Model a utilitzar per l'extracció de l'emoció
        required: true
        schema:
          type: string
          enum:
          - GPT-3.5
          - BERT
          - BETO
          - SVC
      requestBody:
        description: Llista de texts a analitzar
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: array
                  items:
                    required:
                    - id
                    - text
                    type: object
                    properties:
                      id:
                        type: string
                        example: "1"
                      text:
                        type: string
                        example: "Widget not the way it used to be No more 5-day forecast, temp, and local temp."
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    emotion:
                      type: string
                      example: anger
                    text:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "1"
                        text:
                          type: string
                          example: "Widget not the way it used to be No more 5-day forecast, temp, and local temp."
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Bad request error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Unauthorized error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Forbidden error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Not found error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Internal server error"
      x-codegen-request-body-name: Missatges
  /extract-features:
    post:
      summary: Extreure les features d'un conjunt de textos
      description: Retorna un json que conté les features extretes a partir d'un text
      parameters:
      - name: model_features
        in: query
        description: Model a utilitzar per l'extracció de features
        required: true
        schema:
          type: string
          enum:
          - transfeatex
          - t-frex-bert-base-uncased
          - t-frex-bert-large-uncased
          - t-frex-roberta-base
          - t-frex-roberta-large
          - t-frex-xlnet-base-cased
          - t-frex-xlnet-large-cased
      requestBody:
        description: Llista de texts a analitzar
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: array
                  items:
                    required:
                    - id
                    - text
                    type: object
                    properties:
                      id:
                        type: string
                        example: "1"
                      text:
                        type: string
                        example: "Widget not the way it used to be No more 5-day forecast, temp, and local temp."
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    text:
                      type: object
                      properties:
                        features:
                          type: array
                          example: widget not way
                          items:
                            type: string
                        id:
                          type: string
                          example: "1"
                        text:
                          type: string
                          example: "Widget not the way it used to be No more 5-day forecast, temp, and local temp."
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Bad request error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Unauthorized error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Forbidden error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Not found error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Internal server error"
      x-codegen-request-body-name: Missatges
  /analyze-reviews:
    post:
      summary: Analitzar un conjunt de textos
      description: Retorna un json que conté l'emoció o emocions i les features extretes
        a partir d'un text
      parameters:
      - name: model_emotion
        in: query
        description: Model a utilitzar per l'extracció de l'emoció
        required: true
        schema:
          type: string
          enum:
          - GPT-3.5
          - BERT
          - BETO
          - SVC
      - name: model_features
        in: query
        description: Model a utilitzar per l'extracció de features
        required: true
        schema:
          type: string
          enum:
          - transfeatex
          - t-frex-bert-base-uncased
          - t-frex-bert-large-uncased
          - t-frex-roberta-base
          - t-frex-roberta-large
          - t-frex-xlnet-base-cased
          - t-frex-xlnet-large-cased
      requestBody:
        description: Llista de texts a analitzar
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: array
                  items:
                    required:
                    - id
                    - text
                    type: object
                    properties:
                      features:
                        type: array
                        example: widget not way
                        items:
                          type: string
                      id:
                        type: string
                        example: "1"
                      text:
                        type: string
                        example: "Widget not the way it used to be No more 5-day forecast, temp, and local temp."
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    emotion:
                      type: string
                      example: anger
                    features:
                      type: array
                      example:
                      - widget not way
                      items:
                        type: string
                    text:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "1"
                        text:
                          type: string
                          example: "Widget not the way it used to be No more 5-day forecast, temp, and local temp."
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Bad request error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Unauthorized error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Forbidden error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Not found error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                    error:
                        type: string
                        example: "Internal server error"
      x-codegen-request-body-name: Missatges